feat(py): implement identify_partials in terms of regexps since we do not have an AST to walk #90 (#150)
ISSUE: #90

The JS implementation of Handlebars exposes a parse method that
returns an AST, which the `dotprompt.ts` implementation walks
to identify partials in the template. Since, we couldn't find any
methods in the Rust implementation yet that expose this functionality,
we've resorted to using regular expressions to match and find
partials in the template for now.

CHANGELOG:
- [ ] Use regexps to find partials in the template.
- [ ] Add unit tests.
- [ ] Fix some doc comments.